Правки после code review

+++ migrations/000001_create_order_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `order`
 (
     id BINARY(16) NOT NULL,
     status INTEGER NOT NULL,
-    cost FLOAT NOT NULL,
+    cost DECIMAL NOT NULL,
     address VARCHAR (255) NOT NULL,
     created_at DATETIME NOT NULL,
     closed_at DATETIME,
@@ -13,7 +13,7 @@ CREATE TABLE order_item
 (
     order_id BINARY(16) NOT NULL,
     fabric_id BINARY(16) NOT NULL,
-    quantity INT,
+    quantity DECIMAL,
     PRIMARY KEY (order_id, fabric_id),
     FOREIGN KEY (order_id)
         REFERENCES `order`(id)
