+++ pkg/order/application/query/data/order_data.go
@@ -8,7 +8,7 @@ type OrderData struct {
 	ID         string
 	OrderItems []OrderItemData
 	CreatedAt  time.Time
-	Cost       int
+	Cost       float32
 	Status     int
 	Address    string
 }
